teacher-effectiveness/requirements.txt
-e git://github.com/nguyentr17/Insults@b583dbdd326f3de538bccbbf0b13b96e4dbd452d#egg=insults
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
audioread==2.1.8
boto3==1.10.45
botocore==1.13.45
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
dlib==19.19.0
docutils==0.15.2
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
futures==3.1.1
gast==0.2.2
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.8.0
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
importlib-metadata==1.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.30.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.0.2
moviepy==1.0.1
mysqlclient==1.4.6
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
numba==0.46.0
numpy==1.18.0
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.3
pandas-compat==0.1.1
Pillow==6.2.1
plotly==3.6.1
pluggy==0.13.1
proglog==0.1.9
protobuf==3.11.2
pushbullet.py==0.11.0
py==1.8.0
pycparser==2.19
PyMySQL==0.9.3
pyparsing==2.4.5
pyrsistent==0.15.6
pytest==5.3.2
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.19.2
scipy==1.4.1
six==1.13.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.12
statistics==1.0.3.5
tensorflow==1.14.0
termcolor==1.1.0
textblob==0.15.3
tqdm==4.25.0
traitlets==4.3.3
twython==3.7.0
urllib3==1.25.7
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
